+++ src/view/close-popup-button/close-popup-button-template.js
@@ -1,0 +1,5 @@
+export const createClosePopupButtonTemaplte = () => {
+  return `<div class="film-details__close">
+    <button class="film-details__close-btn" type="button">close</button>
+  </div>`;
+};
