+++ src/view/film-card/film-card-template.js
@@ -1,6 +1,4 @@
-import {createElement} from "../utils";
-
-const createFilmCardTemplate = ({id, title, rating, year, duration, genres, poster, description}, commentsCount) => {
+export const createFilmCardTemplate = ({id, title, rating, year, duration, genres, poster, description}, commentsCount) => {
   const [genre] = genres;
   return `<article class="film-card" data-id=${id}>
       <h3 class="film-card__title">${title}</h3>
@@ -20,27 +18,3 @@ const createFilmCardTemplate = ({id, title, rating, year, duration, genres, post
       </div>
     </article>`;
 };
-
-export default class FilmCard {
-  constructor(card, commentsCount) {
-    this._element = null;
-    this._card = card;
-    this._commentsCount = commentsCount;
-  }
-
-  getTemaplate() {
-    return createFilmCardTemplate(this._card, this._commentsCount);
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemaplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
-  }
-}
